pyproject.toml
[tool.poetry]
name = "recharge-to-shopify-subscriptions"
version = "0.0.0"
description = ""
authors = ["Slava Maksimov <hello@slavamak.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
